💄 style: 베스트상품 스켈레톤 UI 추가, 스켈레톤 컴포넌트로 따로뺌
